refactor(effects): rename Light2dEffect → RadialGradientEffect
The shader is a generic procedural radial gradient — solid color at
center fading linearly to transparent at the edge of the host quad.
`WebGLRenderer.drawLight()` is one consumer, but the effect is just as
useful for debug overlays, hotspots, pickup glows, hit-flash indicators,
trigger-zone markers, etc.

Renamed:
- File `effects/light2d.js` → `effects/radialGradient.js`
- Class `Light2dEffect` → `RadialGradientEffect`
- Updated WebGLRenderer.drawLight, the changelog entry, and the
  standalone-API tests.

Documentation reframed around the generic primitive (no Light2d
mention in the description), with five usage examples covering the
main patterns (soft spot, hotspot, pickup highlight, damage flash,
debug overlay).
