+++ packages/melonjs/CHANGELOG.md
@@ -8,9 +8,9 @@
 - Light2d: `illuminationOnly` boolean (default `false`) — when set to `true` the light's own gradient texture isn't drawn, but the light still feeds the cutout pass and the lit-sprite shader's per-frame uniforms. Useful for SpriteIlluminator-style demos where the light should be a logical source, not a visible glow.
 - Light2d: `lightHeight` property (default `max(radiusX, radiusY) * 0.075`) — the Z-axis component of the light direction in the lit shader's `dot(normal, lightDir)`. Low values graze across the surface (dramatic normal-map detail); high values make lighting head-on (more uniform brightness).
 - Two new examples: `Normal Map` (three procedurally-generated 3D orbs in red/green/blue base colors reacting to a moving cursor light, demonstrating that the normal-map controls shape while the color texture controls hue) and `SpriteIlluminator` (faithful port of CodeAndWeb's cocos2d-x dynamic-lighting demo: animated character + foreground prop tile lit by a moving cursor light, full SpriteIlluminator + TexturePacker asset workflow).
-- WebGL: procedural `Light2d` rendering (closes #1430). New `Renderer.drawLight(light)` API replaces the per-light offscreen-canvas pipeline. The WebGL renderer renders each light as a single quad through a shared `Light2dEffect` fragment shader (linear radial falloff, matches the previous Canvas `createRadialGradient` exactly for visual parity); no per-light GL texture is allocated, all lights share one shader program. The Canvas renderer keeps the offscreen-canvas bake but moves it into a renderer-owned `WeakMap<Light2d, target>` cache that auto-invalidates on radius/color/intensity change. Light2d itself becomes pure data — no `CanvasRenderTarget`, no shader knowledge, no renderer reference.
+- WebGL: procedural `Light2d` rendering (closes #1430). New `Renderer.drawLight(light)` API replaces the per-light offscreen-canvas pipeline. The WebGL renderer renders each light as a single quad through a shared `RadialGradientEffect` fragment shader (linear radial falloff, matches the previous Canvas `createRadialGradient` exactly for visual parity); no per-light GL texture is allocated, all lights share one shader program. The Canvas renderer keeps the offscreen-canvas bake but moves it into a renderer-owned `WeakMap<Light2d, target>` cache that auto-invalidates on radius/color/intensity change. Light2d itself becomes pure data — no `CanvasRenderTarget`, no shader knowledge, no renderer reference.
 - Light2d: `resize(radiusX, radiusY)` method. Updates the radii and the underlying bbox so `getBounds()` and `getVisibleArea()` track the new size. Fixes a latent bug where mutating `radiusX/Y` after construction left the rendered light stale while the cutout pass moved.
-- WebGL: new `Light2dEffect` shader effect (`video/webgl/effects/light2d.js`). Standalone-usable: constructor accepts `{ color, intensity, radiusX, radiusY }`, plus `setColor`/`setIntensity`/`setRadii` setters. Used internally by `WebGLRenderer.drawLight`; available for any custom path that wants a procedural soft-circle quad.
+- WebGL: new `RadialGradientEffect` shader effect (`video/webgl/effects/radialGradient.js`). Generic procedural radial gradient — solid color at center fading linearly to transparent at the edge. Constructor accepts `{ color, intensity }`, plus `setColor`/`setIntensity` setters. Used internally by `WebGLRenderer.drawLight`; available for any custom path that wants a soft procedural circle.
 
 ### Changed
 - Lights are now rendered inside the camera's post-effect FBO bracket — vignette, scanlines, ColorMatrix and any other camera shader effect now wrap the lighting output (closes #1398). The `Stage.draw()` lighting block has been removed; rendering happens via the world tree walk and a public `Stage.drawLighting(renderer, camera)` pass invoked by each camera (subclassable for custom lighting).
